use SharedResourcePool. Remove custom wrapper class
